refactor(schedule): extract shared job-service CRUD helpers (phase 5)
Moves the existing shared job verbs out of index.ts into lib/job-service.ts:
guardBakinMutation + READ_ONLY_ERROR + BakinMutationGuard, ensureBakinJob (the
schedule.ensureBakinJob hook handler), getJobByLogicalJobId,
isCronAlreadyGoneError + deleteScheduleJob (+ ScheduleDeleteContext),
readMergedRuntimeJobs, jobToSearchDoc, and indexJob. These already serve both
the REST routes and the exec tools; centralizing them makes the runtime-cron +
search boundary one auditable module and keeps the read-only guard / delete
semantics consistent.

PURE RELOCATION. The route↔exec-tool dedup the audit calls for
(createScheduleJob / applyPauseAction / updateScheduleJob / projectJobDetail)
is deliberately NOT taken here — it folds in a real behavioral drift (the
pause route unconditionally clears pauseUntil while the exec tool keeps it),
so it's left as a separate behavior-touching follow-up rather than silently
collapsed into one behavior. index imports the public surface back and sheds
removeJob / MergedJob / getJobByLogicalJobId imports. 1,060 -> 919.

Verified: typecheck + eslint + full schedule suite (280/0) + full suite
(5124/0) + binary build + DOCKERIZED-RIG E2E against real OpenClaw (isolated):
a full CRUD round-trip — create (indexJob) → PUT then pause (guardBakinMutation,
200) → mutate a bogus id (404 via the guard) → delete (deleteScheduleJob, 200)
→ gone from the list — all moved helpers behaving on the live stack.
